pyproject.toml
[tool.poetry]
name = "haystack-example"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
farm-haystack = {extras = ["ocr"], version = "^1.11.0"}

[tool.poetry.dev-dependencies]

[build-system]
requires = [
    "poetry-core>=1.0.0",
]
build-backend = "poetry.core.masonry.api"
